FEAT: getMyBookList 기능 구현

+++ backend/src/main/java/com/mybrary/backend/domain/book/dto/MyBookGetDto.java
@@ -21,6 +21,6 @@ public class MyBookGetDto {
 
     private Long bookId;
     private String title;
-    private int paperCount;
+    private Long paperCount;
 
 }
